requirements.txt
Pygments==2.15.1
PyYAML==6.0
